Refactor form request unit tests.
- Unit form request tests are now located within the `Sirthxalot\Laravel\I18n\Tests\Http\Requests` namespace.
