+++ tests/fixtures/lang/de_CH/pluralization.php
@@ -1,7 +1,7 @@
 <?php
 
-return [
-    'ago' => '{1} vonere :value minute|[2,*] vor :value minütlis',
-    'apples' => 'Es hät en öpfel|Es git vieli öpfeli mindestens :count',
-    'bananas' => '{0} Es hät kei|[1,19] Es hät es paar|[20,*] Uii sind das viil',
-];
+return array (
+  'ago' => '{1} vonere :value minute|[2,*] vor :value minütlis',
+  'apples' => 'Es hät en öpfel|Es git vieli öpfeli mindestens :count',
+  'bananas' => '{0} Es hät kei|[1,19] Es hät es paar|[20,*] Uii sind das viil',
+);
