+++ tests/fixtures/lang/deu_DE/pluralization.php
@@ -1,7 +1,7 @@
 <?php
 
-return [
-    'ago' => '{1} vor :value Minute|[2,*] vor :value Minuten',
-    'apples' => 'Ein Apfel|Es gibt viele Äpfel mindestens :count',
-    'bananas' => '{0} Es gibt keine|[1,19] Es gibt ein paar|[20,*] Es gibt sehr viele',
-];
+return array (
+  'ago' => '{1} vor :value Minute|[2,*] vor :value Minuten',
+  'apples' => 'Ein Apfel|Es gibt viele Äpfel mindestens :count',
+  'bananas' => '{0} Es gibt keine|[1,19] Es gibt ein paar|[20,*] Es gibt sehr viele',
+);
