+++ tests/fixtures/lang/en/pluralization.php
@@ -1,7 +1,7 @@
 <?php
 
-return [
-    'ago' => '{1} :value minute ago|[2,*] :value minutes ago',
-    'apples' => 'There is one apple|There are many apples at least :count',
-    'bananas' => '{0} There are none|[1,19] There are some|[20,*] There are many',
-];
+return array (
+  'ago' => '{1} :value minute ago|[2,*] :value minutes ago',
+  'apples' => 'There is one apple|There are many apples at least :count',
+  'bananas' => '{0} There are none|[1,19] There are some|[20,*] There are many',
+);
