+++ tests/fixtures/lang/zho/messages.php
@@ -1,8 +1,9 @@
 <?php
 
-return [
-    'nested' => [
-        'message' => '嵌套消息',
-        'welcome' => '你好 :NAME',
-    ],
-];
+return array (
+  'nested' => 
+  array (
+    'message' => '嵌套消息',
+    'welcome' => '你好 :NAME',
+  ),
+);
