+++ tests/fixtures/lang/zho/pluralization.php
@@ -1,7 +1,7 @@
 <?php
 
-return [
-    'ago' => '{1} 一分鐘前|[2,*] :value 幾分鐘前',
-    'apples' => '有一個蘋果|有很多蘋果 :count',
-    'bananas' => '{0} 沒有了|[1,19] 有一些|[20,*] 有許多',
-];
+return array (
+  'ago' => '{1} 一分鐘前|[2,*] :value 幾分鐘前',
+  'apples' => '有一個蘋果|有很多蘋果 :count',
+  'bananas' => '{0} 沒有了|[1,19] 有一些|[20,*] 有許多',
+);
